rework output dict
