+++ environment.yml
@@ -1,5 +1,5 @@
-# conda env create -f environment-dev.yml
-name: pyemittance-dev
+# conda env create -f environment.yml
+name: pyemittance
 channels:
   - conda-forge
 dependencies:
